Move app service node with track1 client to legacy package
